+++ src/lib/app-provider/hooks/useChainId.ts
@@ -1,8 +1,0 @@
-import { useCelatoneApp } from "../contexts";
-
-export const useChainId = () => {
-  const { currentChainId } = useCelatoneApp();
-  if (!currentChainId) throw new Error("Chain ID not found");
-
-  return currentChainId;
-};