+++ src/lib/hooks/useOpenTxTab.ts
@@ -1,15 +1,15 @@
 import { useCallback } from "react";
 
-import { useBaseApiRoute, useChainId } from "lib/app-provider";
+import { useBaseApiRoute, useCelatoneApp } from "lib/app-provider";
 import type { Option } from "lib/types";
 
 export const openNewTab = (url: string) =>
   window.open(url, "_blank", "noopener,noreferrer");
 
 export const useOpenTxTab = (type: "lcd" | "tx-page") => {
-  const chainId = useChainId();
+  const { currentChainId } = useCelatoneApp();
   const txsApiRoute = useBaseApiRoute("txs");
-  const baseUrl = type === "lcd" ? txsApiRoute : `/${chainId}/tx`;
+  const baseUrl = type === "lcd" ? txsApiRoute : `/${currentChainId}/tx`;
 
   return useCallback(
     (txHash: Option<string>) => {