+++ src/lib/services/blockService.ts
@@ -2,7 +2,7 @@ import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
-import { useCelatoneApp, useChainId } from "lib/app-provider";
+import { useCelatoneApp } from "lib/app-provider";
 import {
   getLatestBlockInfoQueryDocument,
   getBlockDetailsByHeightQueryDocument,
@@ -20,7 +20,7 @@ export const useBlocklistQuery = (
   limit: number,
   offset: number
 ): UseQueryResult<BlockInfo[]> => {
-  const chainId = useChainId();
+  const { currentChainId } = useCelatoneApp();
   const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(
     async () =>
@@ -35,7 +35,7 @@ export const useBlocklistQuery = (
               transactions_aggregate,
               validator,
             }) => ({
-              network: chainId,
+              network: currentChainId,
               hash: parseTxHash(hash),
               height,
               timestamp: parseDate(timestamp),
@@ -50,7 +50,7 @@ export const useBlocklistQuery = (
             })
           )
         ),
-    [indexerGraphClient, chainId, limit, offset]
+    [indexerGraphClient, currentChainId, limit, offset]
   );
 
   return useQuery(["blocks", indexerGraphClient, limit, offset], queryFn);
@@ -72,7 +72,7 @@ export const useBlockCountQuery = (): UseQueryResult<number> => {
 export const useBlockDetailsQuery = (
   height: number
 ): UseQueryResult<BlockDetails | null> => {
-  const chainId = useChainId();
+  const { currentChainId } = useCelatoneApp();
   const { indexerGraphClient } = useCelatoneApp();
   const queryFn = useCallback(
     async () =>
@@ -81,7 +81,7 @@ export const useBlockDetailsQuery = (
         .then<BlockDetails | null>(({ blocks_by_pk }) =>
           blocks_by_pk
             ? {
-                network: chainId,
+                network: currentChainId,
                 hash: parseTxHash(blocks_by_pk.hash),
                 height: blocks_by_pk.height,
                 timestamp: parseDate(blocks_by_pk.timestamp),
@@ -99,7 +99,7 @@ export const useBlockDetailsQuery = (
               }
             : null
         ),
-    [indexerGraphClient, chainId, height]
+    [indexerGraphClient, currentChainId, height]
   );
 
   return useQuery(["block_details", indexerGraphClient, height], queryFn, {